refactor: kamoku detail #118
