+++ lib/feature/kamoku_detail/kamoku_detail_page_view.dart
@@ -1,8 +1,9 @@
-import 'package:flutter/material.dart';
-import 'package:dotto/screens/kamoku_detail_syllabus.dart';
-import 'package:dotto/screens/kakomon_list.dart';
+import 'package:dotto/importer.dart';
+
 import 'package:dotto/components/color_fun.dart';
-import 'package:dotto/screens/kamoku_detail_feedback.dart';
+import 'package:dotto/feature/kamoku_detail/kamoku_detail_syllabus.dart';
+import 'package:dotto/feature/kamoku_detail/kamoku_detail_kakomon_list.dart';
+import 'package:dotto/feature/kamoku_detail/kamoku_detail_feedback.dart';
 
 class KamokuDetailPageScreen extends StatefulWidget {
   const KamokuDetailPageScreen(
@@ -94,7 +95,7 @@ class _KamokuDetailPageScreenState extends State<KamokuDetailPageScreen> {
               children: [
                 KamokuDetailSyllabusScreen(lessonId: widget.lessonId),
                 KamokuFeedbackScreen(lessonId: widget.lessonId),
-                KakomonListScreen(
+                KamokuDetailKakomonListScreen(
                     url: widget.kakomonLessonId ?? widget.lessonId),
               ],
             ),
