+++ lib/repository/narrowed_lessons.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:dotto/components/db_config.dart';
+import 'package:dotto/repository/db_config.dart';
 import 'package:dotto/repository/read_json_file.dart';
 import 'package:dotto/components/setting_user_info.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
