+++ lib/screens/home.dart
@@ -1,16 +1,16 @@
-import 'package:dotto/components/db_config.dart';
-import 'package:dotto/screens/kamoku_detail_page_view.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
+import 'package:sqflite/sqflite.dart';
+import 'package:url_launcher/url_launcher.dart';
+
+import 'package:dotto/importer.dart';
 import 'package:dotto/components/color_fun.dart';
+import 'package:dotto/repository/db_config.dart';
+import 'package:dotto/feature/kamoku_detail/kamoku_detail_page_view.dart';
 import 'package:dotto/repository/narrowed_lessons.dart';
 import 'package:dotto/screens/file_viewer.dart';
 import 'package:dotto/screens/course_cancellation.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:sqflite/sqflite.dart';
-import 'package:url_launcher/url_launcher.dart';
 import 'package:dotto/screens/personal_time_table.dart';
-import 'package:intl/intl.dart';
 
 class HomeScreen extends ConsumerStatefulWidget {
   const HomeScreen({super.key});
